Improved handling of termination
Make sure source and sampler are never called after termination.

Also, there was a problem were the sink could be delivered data before it is greeted back, so when it tries to terminate the talkback, it would raise a reference error.
The last test case added in this commit is an example were this probelm would occur.
Reversing the order of subscribing to the puller and greeting back the sink fixed this.